classify note decomposition
